CB-9359: Adds support for .appxbundle creation
.appxbundle files are convenient app packages which can be uploaded to the
Windows Store.  They contain debug or release (typically release) versions
of more than one of the x86, x64, and ARM architectures, and/or they
contain multiple resource packs (for example, both the scale-100 and
scale-240 packs).

To enable bundling, pass in the --bundle flag:

    cordova build windows -- --archs="x86 x64 arm" --bundle

Bundling is disabled if `anycpu` is one of the architectures and multiple
architectures are passed, since bundles can't contain both an
architecture-specific target and a neutral target.  For example, the
following command line would be invalid, and bundling would be disabled:

    cordova build windows -- --archs="x86 x64 arm anycpu" --bundle

Bundling is not supported for Windows 8 as the feature is not available
for that platform in general.

Also corrected an issue in run.spec.js, that caused the build to not be
executed when running as an administrator.

Note: Also contains CR feedback and changed the approach to opt-in instead of opt-out, since
there may be times (specifically around resource generation) in which it
would be interesting to build a .appxbundle with AnyCPU targeted.

This closes #102